% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_matrix.R
\name{as.data.frame.uts_matrix}
\alias{as.data.frame.uts_matrix}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{uts_matrix}(x, ..., method = "wide")
}
\arguments{
\item{x}{a \code{"uts_matrix"} object.}

\item{method}{either \code{"long"} or \code{"wide"}, determining the shape of the output:
\itemize{
  \item \code{"long"}: a \code{data.frame} with one row for each observation for each time series in \code{x}. The \code{data.frame} has four columns denoting the source of each observation (i.e. from which row and column of \code{x} is the observation from?), the observation time, and the observation value.
  \item \code{"wide"}: a \code{data.frame} with one column for each time series in \code{x}. 
}}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Flatten a \code{\link{uts_matrix}} to a \code{\link{data.frame}}.
}
\note{
Only time series with atomic observation values can be coerced to a \code{data.frame}.

This method is helpful for saving a multivariate time series to a human-readable text file.
}
\examples{
as.data.frame(ex_uts_matrix())
as.data.frame(ex_uts_matrix(), method="long")
}
\seealso{
The \code{\link{uts_matrix_long}} and \code{\link{uts_matrix_wide}} constructors provide exactly the opposite funcitonality, i.e. they convert data in \emph{long} and \emph{wide} format, respectively, to a \code{uts_matrix}.
}

