% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.fts.uts_vector}
\alias{as.fts.uts_vector}
\title{Coercion to fts}
\usage{
as.fts.uts_vector(x)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}
}
\value{
An \code{\link[fts]{fts}} object.
}
\description{
Coercion to fts
}
\examples{
if (requireNamespace("fts", quietly = TRUE)) {
  fts::as.fts(ex_uts_vector())
}
}

