% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.uts_vector}
\alias{as.uts_vector}
\alias{as.uts_vector.ts}
\alias{as.uts_vector.fts}
\alias{as.uts_vector.irts}
\alias{as.uts_vector.xts}
\alias{as.uts_vector.zoo}
\title{Coercion to uts_vector}
\usage{
as.uts_vector(x, ...)

\method{as.uts_vector}{ts}(x, ...)

\method{as.uts_vector}{fts}(x, ...)

\method{as.uts_vector}{irts}(x, ...)

\method{as.uts_vector}{xts}(x, ...)

\method{as.uts_vector}{zoo}(x, ...)
}
\arguments{
\item{x}{a time series object of appropriate type.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of class \code{"uts_vector"}.
}
\description{
Convert univariate and multivariate time series objects from other R packages to \code{"uts_vector"} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: convert a \code{\link{ts}} object

\item \code{fts}: convert an \code{\link[fts]{fts}} object

\item \code{irts}: convert an \code{\link[tseries]{irts}} object

\item \code{xts}: convert a \code{\link[xts]{xts}} object

\item \code{zoo}: convert a \code{\link[zoo]{zoo}} object
}}

\examples{
# Convert a bivariate "ts"
ts1 <- ts(matrix(1:20, 10, 2), start=c(2016, 1), frequency=12, names=c("apples", "oranges"))
utsv <- as.uts_vector(ts1)
utsv[[1]]
utsv$oranges

# Convert a 2-dimensional "fts" and remove NA observations
if (requireNamespace("fts", quietly = TRUE)) {
  x <- fts::fts(index=seq(from=Sys.Date(), by="months", length.out=24),data=1:24)
  y <- fts::fts(index=seq(from=Sys.Date(), by="months", length.out=12),data=13:24)
  fts1 <- cbind(x, y)
  na.omit(as.uts_vector(fts1))
}

# Convert a 3-dimensional "irts"
if (requireNamespace("tseries", quietly = TRUE)) {
  irts1 <- tseries::irts(as.POSIXct("2015-01-01") + days(c(1, 3, 7, 9)), matrix(1:12, 4))
  as.uts_vector(irts1)
}

# Convert a 3-dimensional "its"
if (requireNamespace("its", quietly = TRUE)) {
  mat <- matrix(1:6, nrow=2)
  colnames(mat) <- c("a", "b", "c")
  rownames(mat) <- c("2003-01-01","2003-01-04")
  its1 <- its::its(mat)
  as.uts_vector(its1)
}

# Convert a 3-dimensional "xts"
if (requireNamespace("zoo", quietly = TRUE)) {
  xts1 <- xts::xts(matrix(1:12, 4, 3), as.Date("2003-01-01") + 0:3)
  as.uts_vector(xts1)
}

# Convert a 3-dimensional "zoo"
if (requireNamespace("zoo", quietly = TRUE)) {
  zoo1 <- zoo::zoo(matrix(1:12, 4, 3), as.Date("2003-01-01") + 0:3)
  utsv <- as.uts_vector(zoo1)
  utsv[[1]]
}

}
\seealso{
\code{\link[uts]{as.uts}} for converting just univariate time series.
}
