% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.xts.uts_vector}
\alias{as.xts.uts_vector}
\title{Coercion to xts}
\usage{
as.xts.uts_vector(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{\link[xts]{xts}} object.
}
\description{
Coercion to xts
}
\examples{
if (requireNamespace("xts", quietly = TRUE)) {
  xts::as.xts(ex_uts_vector())
}
}

