% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.zoo.uts_vector}
\alias{as.zoo.uts_vector}
\title{Coercion to zoo}
\usage{
as.zoo.uts_vector(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{\link[zoo]{zoo}} object.
}
\description{
Coercion to zoo
}
\examples{
if (requireNamespace("zoo", quietly = TRUE)) {
  # Bivariate uts_vector with synchronized observation times
  utsv <- c(a=ex_uts(), b=ex_uts() + 3)
  zoo::as.zoo(utsv)
  
  # # Bivariate uts_vector with non-synchronized observation times
  zoo::as.zoo(ex_uts_vector())
}
}
