% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_ts.R
\name{ex_uts_vector}
\alias{ex_uts_vector}
\alias{ex_uts_vector2}
\title{Example Multivariate Time Series}
\usage{
ex_uts_vector()

ex_uts_vector2()
}
\value{
\code{ex_uts_vector()} returns a numeric \code{"uts_vector"} with two time series.

\code{ex_uts2()} returns a non-numeric \code{"uts_vector"} with two time series.
}
\description{
Create time series vectors that can be used for code examples and testing.
}
\examples{
ex_uts_vector()
ex_uts_vector2()
}

