% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first.uts_vector}
\alias{first.uts_vector}
\alias{last.uts_vector}
\title{First and Last Observation Values}
\usage{
\method{first}{uts_vector}(x, ...)

\method{last}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{first()} returns the first observation value of each time series.

\code{last()} returns the last observation value of each time series.
}
\description{
Get the first and last observation value, respectively, of each time series.
}
\examples{
first(ex_uts_vector())
last(ex_uts_vector2())

first(ex_uts_matrix())
}

