% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs.R
\name{is.na.uts_vector}
\alias{is.na.uts_vector}
\title{Not Available / Missing Values}
\usage{
\method{is.na}{uts_vector}(x)
}
\arguments{
\item{x}{a \code{"uts_vector"}, \code{"uts_matrix"}, or \code{"uts_data_frame"} object.}
}
\value{
An object of the same class as \code{x}, but with observation values replaced with \code{TRUE} or \code{FALSE}, indicating which observation values of \code{x} are \code{NA}.
}
\description{
Find missing values.
}
\examples{
# Set observation to NA
test <- ex_uts_vector()
test$oranges$values[c(2, 4)] <- NA

# Get logical "uts_vector", indicating which observations are NA
is.na(test)
is.na(test)$oranges
}
\seealso{
\code{\link{na.omit}}

\code{\link[uts:is.na.uts]{is.na}}, \code{\link[uts:na.omit.uts]{na.omit}} for \code{"uts"} objects.

\code{\link[base:is.na]{is.na}}, \code{\link[stats:na.fail]{na.omit}} in base \R.
}

