% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{is.uts_vector}
\alias{is.uts_vector}
\title{Is Object a uts_vector?}
\usage{
is.uts_vector(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Return \code{TRUE} if and only if the argument is a \code{"uts_vector"} object.
}
\examples{
is.uts_vector(uts_vector())
is.uts_vector(ex_uts_vector())
is.uts_vector(5)
}
\keyword{internal}

