% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Math.uts_vector}
\alias{Math.uts_vector}
\title{Math Group Methods for uts_vector}
\usage{
\method{Math}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply the \code{\link{Math}} methods in base \R{} to the observation values of \code{"uts_vector"} objects.
}
\examples{
# Take the base-2 logarithm of the observation values and return the corresponding "uts_vector"
log(ex_uts_vector(), base=2)

# Calculate the cumulative sum of the observation values and return the corresponding "uts_vector"
cumsum(ex_uts_vector())
}
\seealso{
\code{\link{groupGeneric}}
}

