% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{mean.uts_vector}
\alias{mean.uts_matrix}
\alias{mean.uts_vector}
\title{Mean, Median, and Standard Deviation of Observation Values}
\usage{
\method{mean}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Calculate the mean, median, and standard deviation, respectively, of the observation values of a multivariate time series.
}
\examples{
mean(ex_uts_vector())
mean(ex_uts_matrix())
}
\seealso{
\code{\link[uts]{mean.uts}}, \code{\link[uts]{median.uts}}, \code{\link[uts]{sd.uts}}

\code{\link{mean}}, \code{\link{median}}, \code{\link{sd}} in base \R.
}
\keyword{internal}

