% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs.R
\name{na.omit.uts_vector}
\alias{na.omit.uts_matrix}
\alias{na.omit.uts_vector}
\title{Remove NAs}
\usage{
\method{na.omit}{uts_vector}(object, ...)
}
\arguments{
\item{object}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Returns the object with incomplete cases removed.
}
\examples{
# Remove NAs from a "uts_vector"
test <- ex_uts_vector()
test$oranges$values[c(2, 4)] <- NA
na.omit(test)
}
\seealso{
\code{\link[uts]{is.na.uts}}, \code{\link[uts]{na.omit.uts}} for \code{"uts"} objects.

\code{\link{is.na}}, \code{\link{na.omit}} in base \R.
}

