% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_single_uts_vector}
\alias{plot_single_uts_vector}
\title{Plot a uts_vector in a single plot}
\usage{
plot_single_uts_vector(
  x,
  ...,
  max_dt = ddays(Inf),
  xlab = "",
  ylab = "",
  col = seq_along(x),
  lty = 1,
  lwd = 1,
  pch = 1,
  type = "l",
  legend = TRUE,
  legend.x = "topright",
  legend.y = NULL
)
}
\arguments{
\item{x}{a \code{"uts_vector"} object with numeric or logical observation values.}

\item{\dots}{arguments passed to \code{\link[uts]{plot.uts}}.}

\item{max_dt}{a non-negative \code{\link[lubridate]{duration}} object. Consecutive observations that are more than this amount apart in time, are not connected by a line in the graph.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis}

\item{col, lty, lwd, pch, type}{graphical parameters. See \code{\link{plot.default}}.}

\item{legend}{boolean. Whether to add a legend to the plot.}

\item{legend.x, legend.y}{the x and y co-ordinates to be used to position the legend.}
}
\description{
A helper function that implements \code{plot.uts_vector} for argument \code{plot.type="single"}.
}
\examples{
plot_single_uts_vector(ex_uts_vector(), xlab="time")
plot_single_uts_vector(ex_uts_vector(), type="o", main="Fruit", max_dt=dhours(12))
plot_single_uts_vector(ex_uts_vector(), type="p", pch=2, ylim=c(40, 60), cex=2)
}
\seealso{
\code{\link{matplot}}
}
\keyword{internal}
