% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{print.uts_vector}
\alias{print.uts_vector}
\title{Print Values}
\usage{
\method{print}{uts_vector}(x)
}
\arguments{
\item{x}{an object of class \code{"uts_vector"}.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Print Values
}
\examples{
uts_vector()
print(uts_vector(a=ex_uts(), b=ex_uts2()))
}
\seealso{
\code{\link[base:print]{print}}
}
\keyword{internal}

