% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{rep.uts}
\alias{rep.uts}
\alias{rep.uts_vector}
\title{Repeat uts and uts_vector}
\usage{
\method{rep}{uts}(x, ...)

\method{rep}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} or \code{"uts_vector"}.}

\item{\dots}{further arguments passed to \code{\link{rep}} in base \R.}
}
\value{
An object of class \code{"uts_vector"}.
}
\description{
Create a \code{"uts_vector"} by replicating the individual \code{"uts"} of the input \code{x}.
}
\details{
This method is a wrapper around \code{\link{rep}} in base \R that makes sure the returned object is of class \code{"uts_vector"}.
}
\examples{
# Repeat "uts"
rep(ex_uts(), 4)

# Repeat "uts_vector"
rep(ex_uts_vector(), times=3)
rep(ex_uts_vector(), each=3)
}
