% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply.R
\name{sapply}
\alias{sapply}
\alias{sapply.default}
\title{Generic sapply function}
\usage{
sapply(X, ...)

\method{sapply}{default}(X, ...)
}
\arguments{
\item{X}{an \R object.}

\item{further}{arguments passed to or from methods.}
}
\description{
The function is needed, because \code{\link[base:sapply]{sapply}} of base \R is not generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: simply calls the default implementation of base \R
}}
\note{
As recommended in Section 7.1 ("Adding new generics") of "Writing R Extensions", the implementation of \code{\link{sapply.default}} has been made a wrapper around \code{\link[base:sapply]{base::sapply}}.
}
\keyword{internal}

