% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply.R
\name{sapply.uts_vector}
\alias{sapply.uts_vector}
\title{Apply a Function over a Multivariate Time Series}
\usage{
\method{sapply}{uts_vector}(X, FUN, ...)
}
\arguments{
\item{X}{a \code{"uts_vector"} object.}

\item{FUN}{the function to be applied to each element of \code{X}.}

\item{\ldots}{optional arguments to \code{FUN}.}
}
\description{
Apply a Function over a Multivariate Time Series
}
\examples{
sapply(ex_uts_vector(), length)
sapply(ex_uts_vector(), range)
sapply(ex_uts_vector(), log)
}

