% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{start.uts_vector}
\alias{end.uts_vector}
\alias{start.uts_vector}
\title{First and Last Observation Times}
\usage{
\method{start}{uts_vector}(x, ...)

\method{end}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{end()} returns the first observation time of each time series.

\code{end()} returns the last observation time of each time series.
}
\description{
Get the first and last observation time, respectively, of each time series.
}
\examples{
start(ex_uts_vector())
end(ex_uts_vector2())
}

