% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{[.uts_vector}
\alias{[.uts_vector}
\title{Extract or Replace Parts of a uts_vector}
\usage{
\method{[}{uts_vector}(x, i, j, drop = TRUE, ...)
}
\arguments{
\item{x}{a \code{"uts)vector"} object.}

\item{i}{either a strictly increasing sequence of \code{\link{POSIXct}} date-times, or a \code{"uts"} or \code{"uts_vector"} with \code{\link{logical}} observation values.}

\item{j}{index specifying the time series to extract or replace. The index can be a \code{numeric} or \code{character} vector or empty (missing) or \code{NULL}. Numeric values are coerced to integer as by \code{\link{as.integer}} (and hence truncated towards zero). Character vectors will be matched to the \code{\link{names}} of the object.}

\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible dimension.}

\item{\dots}{further arguments passed to \code{\link[uts]{sample_values.uts}}.}
}
\description{
The accessor method (\code{"["}) extracts either (i) a sub-sampled time series vector, (ii) a subset of the time series vector, or (iii) both. The replacement method (\code{"[<-"}) inserts observation values at the provided observation times, replacing observations values for already existing observation times (if any).
}
\examples{
# Extract subset of time series vector
x <- ex_uts_vector()
x[, 1]
x[, "oranges", drop=FALSE]
x[, 2:1]

# Sample at multiple time points, optionally restricting the maximum time difference
# between sampling and observation times
x <- ex_uts_vector()
times <- as.POSIXct(c("2007-11-08 11:01:00", "2007-11-09 15:16:00"))
x[times]
x[times, interpolation="linear"]

# Both at the same time, i.e. subsampling and subsetting
x[times, "oranges"]
x[times, c(1, 2, 1)]
}

