% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_matrix.R
\name{uts_matrix_wide}
\alias{uts_matrix_wide}
\title{Create uts_matrix from wide tabular data}
\usage{
uts_matrix_wide(values, times, names = base::rownames(values),
  fields = base::colnames(values))
}
\arguments{
\item{values}{a matrix or data.frame of observation values.}

\item{times}{a \code{\link{POSIXct}} object. The observation time for each row of \code{values}.}

\item{names}{a character vector. The the row name (also known as \emph{record}) of each row of \code{values}.}

\item{fields}{a character vector. The the column names (also known as \emph{field}) of each column of \code{values}.}
}
\value{
An object of class \code{"uts_matrix"}. The number of rows is given by to the number of records (distinct \code{rownames}), while the number of columns is given by the number of fields (distinct \code{colnames}).
}
\description{
Create a \code{"uts_matrix"} from \emph{wide} tabular data (see \href{https://en.wikipedia.org/wiki/Wide_and_narrow_data}{Wikipedia}). For data in this format, each row is a vector of observations (also known as \emph{record}) at specific time point.
}
\examples{
values <- matrix(1:15, 5, 3)
rownames(values) <- c("A", "A", "B", "B", "C")
colnames(values) <- c("a", "b", "c")
times <- as.POSIXct("2016-01-01") + days(1:5)
uts_matrix_wide(values, times)

# Same, but with manually provided entity and attribute names
uts_matrix_wide(values, times, names=c("CA", "CA", "FR", "US", "US"),
  fields=c("apple", "banana", "orange"))
}
\keyword{classes}
\keyword{ts}

