% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{uts_vector}
\alias{c.uts}
\alias{c.uts_vector}
\alias{uts_vector}
\title{Unevenly-spaced Time Series Vector}
\usage{
uts_vector(...)

\method{c}{uts}(...)

\method{c}{uts_vector}(...)
}
\arguments{
\item{\dots}{zero or more \code{\link{uts}} objects.}
}
\value{
An object of class \code{"uts_vector"}.
}
\description{
Create a vector of unevenly spaced time series (\code{"uts_vector"}).
}
\section{Methods (by generic)}{
\itemize{
\item \code{c}: constructor for \code{"uts_vector"} object out of \code{"uts"} objects.

\item \code{c}: constructor for \code{"uts_vector"} object out of \code{"uts"} and other \code{"uts_vector"} objects.
}}
\note{
A virtual class \code{"uts_virtual"} exists from which \code{"uts"}, \code{"uts_vector"}, \code{"uts_matrix"}, and \code{"uts_data_frame"} inherit: it is used to allow operations such as subtraction to mix the classes.
}
\examples{
# Two equivalent constructors
uts_vector(apples=ex_uts(), oranges=ex_uts2())
c(apples=ex_uts(), oranges=ex_uts2())

# Create "uts_vector" out of both "uts" and "uts_vector" objects
c(ex_uts(), ex_uts_vector(), kiwis=ex_uts2())

# Empty "uts_vector"
uts_vector()
}
\seealso{
\code{\link{rep.uts}}, \code{\link{uts_vector_wide}}, \code{uts_vector_long} for alternative constructors.
}
\keyword{classes}
\keyword{ts}

