% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{uts_vector_long}
\alias{uts_vector_long}
\title{Create uts_vector from long tabular data}
\usage{
uts_vector_long(values, times, names)
}
\arguments{
\item{values}{a vector observation values.}

\item{times}{a \code{\link{POSIXct}} object. The matching observation times.}

\item{names}{a character vector. Thehe source of each observation.}
}
\value{
An object of class \code{"uts_vector"} with length given by to the number of distinct \code{names}.
}
\description{
Create a \code{"uts_vector"} from \emph{long} (also known as \emph{narrow}) tabular data. Data in this format has three different columns; the observation values, the observation times, and the source of each observation.
}
\examples{
uts_vector_long(values=1:10, times=as.POSIXct("2010-01-01") + days(1:10), names=rep(c("a", "b", "c"), length=10))
}
\seealso{
\code{\link{uts_vector_wide}}
}
\keyword{classes}
\keyword{ts}

