% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{window.uts_vector}
\alias{time.uts_matrix}
\alias{window.uts_vector}
\title{Time Window}
\usage{
\method{window}{uts_vector}(x, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{start, end}{\code{\link{POSIXct}} object or coercible using \code{\link{as.POSIXct}}. The start and end times, respectively, for the individual subperiod time series. If there are fewer times than time series, then they are recycled in the standard fashion.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Extract a subperiod time series between times \code{start} and \code{end}.
}
\examples{
# For each time series, drop observations before 2007-11-09 Eastern Standard Time
window(ex_uts_vector(), start="2007-11-09 EST")

# Use a different end time for each subperiod time series
window(ex_uts_vector(), end=c("2007-11-09 12:00:00 EST", "2007-11-09 EST"))
}
\seealso{
\code{\link{head}}, \code{\link{head_t}}, \code{\link{tail}}, \code{\link{tail_t}} for other methods that extract a subperiod time series.
}

