% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_compatability.R
\name{compatability}
\alias{compatability}
\alias{cycle.uts_vector}
\alias{frequency.uts_vector}
\title{Compatability with ts class}
\usage{
\method{frequency}{uts_vector}(x, ...)

\method{cycle}{uts_vector}(x, ...)
}
\arguments{
\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{cycle()} and \code{frequency()} give an error message, because  \code{"uts_vector"}, \code{"uts_matrix"}, and \code{"uts_data_frame"} objects objects, by definition, do not have a fixed number of observations in a given time interval.
}
\description{
These methods exist solely to ensure that methods intended for \code{"ts"} objects in base \R are not accidentally applied to \code{"uts_vector"}, \code{"uts_matrix"}, and \code{"uts_data_frame"} objects.
}
\examples{
\dontrun{frequency(ex_uts_vector())}
\dontrun{cycle(ex_uts_vector())}
}
\seealso{
\code{\link{ts}}
}
\keyword{internal}

