% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ts.R
\name{ex_uts_vector}
\alias{ex_uts_vector}
\alias{ex_uts_vector2}
\alias{ex_uts_matrix}
\title{Example Multivariate Time Series}
\usage{
ex_uts_vector()

ex_uts_vector2()

ex_uts_matrix()
}
\value{
\code{ex_uts_vector()} returns a numeric \code{"uts_vector"} with two time series.

\code{ex_uts_vector2()} returns a non-numeric \code{"uts_vector"} with two time series.

\code{ex_uts_matrix()} returns a two-by-two numeric \code{"uts_matrixr"}.
}
\description{
Create time series vectors that can be used for code examples and testing.
}
\examples{
ex_uts_vector()
ex_uts_vector2()
ex_uts_matrix()
}
