% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NAs.R
\name{na.omit.uts_vector}
\alias{na.omit.uts_vector}
\title{Remove NAs}
\usage{
\method{na.omit}{uts_vector}(object, ...)
}
\arguments{
\item{object}{a \code{"uts_vector"}, \code{"uts_matrix"}, or \code{"uts_data_frame"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Returns the object with incomplete cases removed.
}
\examples{
# Remove NAs from a "uts_vector"
tmp <- ex_uts_vector()
tmp$oranges$values[c(2, 4)] <- NA
na.omit(tmp)
}
\seealso{
\code{\link[base:is.na]{is.na}}, \code{\link[stats:na.fail]{na.fail}}, \code{\link[stats:na.fail]{na.omit}} in base \R.

\code{\link[uts:is.na.uts]{is.na}} \code{\link[uts:na.omit.uts]{na.omit}} for \code{"uts"} objects.
}

