% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\name{Ops_uts_vector}
\alias{Ops_uts_vector}
\title{Ops Group Methods for uts_vector}
\usage{
Ops_uts_vector(x, x2, .Generic)
}
\description{
Ops Group Methods for uts_vector
}
\examples{
# Unary operators
-ex_uts()
!(ex_uts() > 48)

# Binary operators
ex_uts() + ex_uts()
ex_uts() + 20
20 + ex_uts()
}
\keyword{internal}

