% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.uts_vector}
\alias{plot.uts_vector}
\title{Plot a uts_vector}
\usage{
\method{plot}{uts_vector}(x, ..., plot.type = "multiple",
  ask = getOption("device.ask.default"), legend = TRUE,
  legend.x = "topright", legend.y = NULL)
}
\arguments{
\item{x}{a \code{"uts_vector"} object with numeric or logical observation values.}

\item{plot.type}{\code{"multiple"} plots the time eries on multiple plots, while \code{"single"} superimposes them on a single plot}

\item{ask}{logical. If \code{TRUE} (and the \R session is interactive) the user is asked for input, before a new figure is drawn.}

\item{legend}{boolean. For \code{plot.type="single"}, whether to add a \code{\link{legend}} to the plot.}

\item{legend.x, legend.y}{the x and y co-ordinates to be used to position the legend. See \code{\link{legend}}.}

\item{\dots}{arguments passed to \code{\link[uts]{plot.uts}}.}
}
\description{
Plot the individual time series of a \code{"uts_vector"} using \code{\link[uts]{plot.uts}}.
}
\examples{
# Plot all time series at once
old <- par(mfrow=c(1, 2))
plot(ex_uts_vector())
plot(ex_uts_vector(), max_dt=dhours(12), type="b")   # don't connect points more than 12 hours apart
par(old)

# Plot one time series at a time
plot(ex_uts_vector(), ask=TRUE)

# Plot multiple time series in a single plot by superimposing the plots
plot(ex_uts_vector(), plot.type="single")
}
\seealso{
\code{\link{plot.uts}}
}

