% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.uts_vector}
\alias{plot.uts_vector}
\title{Plot a uts_vector}
\usage{
\method{plot}{uts_vector}(x, ..., ask = getOption("device.ask.default"))
}
\arguments{
\item{x}{a \code{"uts_vector"} object with numeric or logical observation values.}

\item{ask}{logical. If \code{TRUE} (and the \R session is interactive) the user is asked for input, before a new figure is drawn.}

\item{\dots}{arguments passed to \code{\link[uts]{plot.uts}}.}
}
\description{
Plot the individual time series of a \code{"uts_vector"} using \code{\link[uts]{plot.uts}}.
}
\examples{
# plot all time series at once
old <- par(mfrow=c(1, 2))
plot(ex_uts_vector())
plot(ex_uts_vector(), max_dt=dhours(12), type="b")   # don't connect points more than 12 hours apart
par(old)

# plot one time series at a time
plot(ex_uts_vector(), ask=TRUE)
}

