% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_single_uts_vector}
\alias{plot_single_uts_vector}
\title{Plot a uts_vector in a single plot}
\usage{
plot_single_uts_vector(x, ..., xlab = "", ylab = "", legend = TRUE,
  legend.x = "topright", legend.y = NULL)
}
\arguments{
\item{x}{a \code{"uts_vector"} object with numeric or logical observation values.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis}

\item{legend}{boolean. Whether to add a legend to the plot.}

\item{legend.x, legend.y}{the x and y co-ordinates to be used to position the legend.}

\item{\dots}{arguments passed to \code{\link[uts]{plot.uts}}.}
}
\description{
A helper function that implements \code{plot.uts_vector} for argument \code{plot.type="single"}.
}
\examples{
plot_single_uts_vector(ex_uts_vector())
}
\seealso{
\code{\link{matplot}}
}
\keyword{internal}

