% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.uts_vector}
\alias{print.uts_matrix}
\alias{print.uts_vector}
\title{Print Multivariate Time Series}
\usage{
\method{print}{uts_vector}(x, ...)

\method{print}{uts_matrix}(x, print_times = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{print_times}{boolean. For two-dimensional time series objects, whether to print the start and end times of the individual time series.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Print basic information (the number of data points, and the first and last observation time) about the individual \code{"uts"} in a multivariate time series.
}
\examples{
# Print "uts_vector"
uts_vector()
print(uts_vector(a=ex_uts(), b=ex_uts2()))

# Print "uts_matrix"
ex_uts_matrix()
print(ex_uts_matrix(), print_times=TRUE)

# Print "uts_data_frame"
}
\seealso{
\code{\link[base:print]{print}} in base \R.
}
\keyword{internal}

