% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_values.uts_vector}
\alias{sample_values.uts_vector}
\title{Sample Values}
\usage{
\method{sample_values}{uts_vector}(x, time_points, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{"uts_vector"} object where each individual \code{"uts"} as atomic observation values.}

\item{time_points}{a strictly increasing sequence of \code{\link{POSIXct}} date-times.}

\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible dimension. See the examples and \code{\link{drop}}.}

\item{\dots}{arguments passed to \code{\link[uts]{sample_values.uts}} for sampling from the individual time series.}
}
\value{
A \code{matrix} if all sampled observation values are of the same \code{\link{type}}, and a \code{data.frame} otherwise. However, when using argument \code{drop=TRUE} the result may be simplified further.
}
\description{
Sample each time series in a \code{"uts_vector"} at given sampling times and combine the sampled values into a \code{matrix} or \code{data.frame}.
}
\examples{
times <- as.POSIXct(c("2007-11-09", "2007-11-10"))

# Sample with last-point and linear interpolation
sample_values(ex_uts_vector(), times)
sample_values(ex_uts_vector(), times, interpolation="linear")

# Sample with and without dropping of length-one dimensions
sample_values(ex_uts_vector(), times[1])
sample_values(ex_uts_vector(), times[1], drop=FALSE)

# Store sampled values in data.frame if of different type
x <- ex_uts()
y <- uts(letters[1:6], x$times)
utsv <- c(x, y)
sample_values(utsv, times)

# Error, because not all time series have atomic observation values
\dontrun{sample_values(ex_uts_vector2(), times)}
}

