% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{sapply.uts_vector}
\alias{sapply.uts_vector}
\title{Apply a Function over a Multivariate Time Series}
\usage{
\method{sapply}{uts_vector}(X, ...)
}
\arguments{
\item{X}{a \code{"uts_vector"} object.}

\item{\ldots}{arguments passed to \code{sapply} in base \R.}
}
\description{
This method is a wrapper around \code{\link[base:sapply]{sapply}} in base \R. If possible, it further simplifies the output to a \code{"uts_vector"}, \code{"uts_matrix"}, or \code{"uts_data_frame"}.
}
\examples{
# Same results as sapply() in base R
sapply(ex_uts_vector(), length)
sapply(ex_uts_vector(), range)

# Results that are further simplified to a "uts_vector"
sapply(ex_uts_vector(), log)
sapply(ex_uts_vector2(), lag_t, ddays(1))
}

