% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{sapply.uts_vector}
\alias{sapply.uts_matrix}
\alias{sapply.uts_vector}
\title{Apply a Function over a Multivariate Time Series}
\usage{
\method{sapply}{uts_vector}(X, ...)

\method{sapply}{uts_matrix}(X, ...)
}
\arguments{
\item{X}{a \code{"uts_vector"} or \code{"uts_amtrixr"} object.}

\item{\ldots}{arguments passed to \code{sapply} in base \R.}
}
\description{
A wrapper around \code{\link[base:sapply]{sapply}} in base \R. If possible, it further simplifies the output to a \code{"uts_vector"} or \code{"uts_matrix"}.
}
\examples{
# Same results as sapply() in base R
sapply(ex_uts_vector(), length)
sapply(ex_uts_vector(), range)

# Results that are further simplified to a "uts_vector"
sapply(ex_uts_vector(), log)
sapply(ex_uts_vector2(), lag_t, ddays(1))

# Results that are further simplified to a matrix "uts_matrix"
sapply(ex_uts_matrix(), length)
sapply(ex_uts_matrix(), sqrt)
sapply(ex_uts_matrix(), function(x) "a")
}

