% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end.R
\name{start.uts_vector}
\alias{end.uts_vector}
\alias{start.uts_vector}
\title{First and Last Observation Times}
\usage{
\method{start}{uts_vector}(x, ...)

\method{end}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{start()} returns the first observation time of each time series as a \code{\link{POSIXct}} object. The time zone is determined by the first time series.

\code{end()} returns the last observation time of each time series as a \code{\link{POSIXct}} object. The time zone is determined by the first time series.
}
\description{
Get the first and last observation time, respectively, of each time series.
}
\examples{
start(ex_uts_vector())
end(ex_uts_vector2())

# Return vector, because POSIXct objects cannot be stored in a matrix
start(ex_uts_matrix())
}

