% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{[.uts_vector}
\alias{[.uts_vector}
\alias{[<-.uts_vector}
\title{Extract or Replace Parts of a uts_vector}
\usage{
\method{[}{uts_vector}(x, i, j, drop = TRUE, ...)

\method{[}{uts_vector}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{i}{either a strictly increasing sequence of \code{\link{POSIXct}} date-times, or a \code{"uts"} or \code{"uts_vector"} with \code{\link{logical}} observation values.}

\item{j}{index specifying the time series to extract or replace. The index can be a \code{numeric} or \code{character} vector or empty (missing) or \code{NULL}. Numeric values are coerced to integer as by \code{\link{as.integer}} (and hence truncated towards zero). Character vectors will be matched to the \code{\link{names}} of the object.}

\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible dimension.}

\item{value}{a vector of observation values, a \code{"uts"}, a or \code{"uts_vector"}.}

\item{\dots}{further arguments passed to \code{\link[uts]{sample_values.uts}}.}
}
\description{
The accessor method (\code{"["}) extracts either (i) a sub-sampled time series vector, (ii) a subset of the time series vector, or (iii) both. The replacement method (\code{"[<-"}) inserts observation values at the provided observation times, replacing observations values for already existing observation times (if any).
}
\examples{
##############################
# Subsampling and Subsetting #
##############################

# Extract subset of time series vector
x <- ex_uts_vector()
x[, 1]
x[, "oranges", drop=FALSE]
x[, 2:1]

# Subsampling using a POSIXct vector
x <- ex_uts_vector()
times <- as.POSIXct(c("2007-11-08 11:01:00", "2007-11-09 15:16:00"))
x[times]
x[times, interpolation="linear"]

# Subsampling using a "uts" with logical observation values
x <- ex_uts_vector()
x[ex_uts() > 48]

# Subsampling and subsetting at the same time
# 1.) using separate subsampling and subsetting indicies
x[times, "oranges"]
x[times, c(1, 2, 1)]
# 2.) using "uts_vector" with logical observation values
x[x > 48]


###############################
# Replacements and Insertions #
###############################

# Replace subset of time series vector with a single time series
x <- ex_uts_vector()
x[, "oranges"] <- uts(values=50, times=Sys.time())
x[, "nuts"] <- head(ex_uts(), 2)
x[, "apples"] <- NULL

# Same, but use "uts_vector" for replacing
x <- c(ex_uts_vector(), nuts=ex_uts())
x[, 1:2] <- c(uts(), ex_uts())
x[, c("apples", "oranges")] <- NULL

# Replace/insert observations of single time series
# 1.) Insert single value into the first time series 
# 2.) Insert multiple values into the time series named "oranges"
x <- ex_uts_vector()
x[Sys.time(), 1] <- 50
x[Sys.time() + ddays(1:2), "oranges"] <- c(52, 53)

# Replace/insert observations of multiple time series
# 1.) Insert single value into every time series
# 2.) Insert the same multiple values into every time series
x <- ex_uts_vector()
x[Sys.time(), ] <- 51
x[Sys.time() + ddays(1:2), ] <- c(52, 53)
}

