% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Summary.uts_vector}
\alias{Summary.uts_vector}
\title{Summary Group Methods for uts_vector}
\usage{
\method{Summary}{uts_vector}(x, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply the \code{\link{Summary}} methods in base \R{} to the time series of \code{"uts_vector"} objects.
}
\examples{
# Get the smallest observation value of each time, ignoring NAs
min(ex_uts_vector(), na.rm=TRUE)

# Check which time series have any observation value larger than 49
any(ex_uts_vector() > 49)
}
\seealso{
\code{\link{groupGeneric}}, \code{\link[uts]{Summary.uts}}
}

