% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{summary.uts_vector}
\alias{summary.uts_vector}
\alias{summary.uts_matrix}
\title{Summary of Time Series Values}
\usage{
\method{summary}{uts_vector}(object, ...)

\method{summary}{uts_matrix}(object, ...)
}
\arguments{
\item{object}{a \code{"uts_vector"} or \code{"uts_matrix"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply \code{\link{summary}} from base \R to the observation values of a multivariate time series.
}
\note{
This method primarily exists because \code{\link{summary.default}} produces an error message.
}
\examples{
summary(ex_uts_vector())
summary(ex_uts_matrix())
}
\seealso{
\code{\link[uts]{summary.uts}}
}
\keyword{internal}
