% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{time.uts_vector}
\alias{time.uts_vector}
\title{Observation Times}
\usage{
\method{time}{uts_vector}(x, tolerance = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{a \code{"uts_vector"} object.}

\item{tolerance}{a non-negative number, indicating the tolerance for numerical noise in the observation times. Observation times less than this threshold apart are treated as identical.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Get the sorted union of observation times of a \code{"uts_vector"} object.
}
\examples{
time(ex_uts_vector())
time(ex_uts_vector2())
}
\seealso{
\code{\link[uts]{time.uts}}, \code{\link[uts]{sorted_union}}
}
