% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_long.R
\name{uts_matrix_long}
\alias{uts_matrix_long}
\title{Create uts_matrix from long tabular data}
\usage{
uts_matrix_long(values, times, names = base::names(values), fields)
}
\arguments{
\item{values}{a vector observation values.}

\item{times}{a \code{\link{POSIXct}} object. The matching observation times.}

\item{names}{a character vector. The the matching entity names for the observations. By default, the names of \code{values} are used.}

\item{fields}{a character vector. The the matching attribute/field names for the observations.}
}
\value{
An object of class \code{"uts_matrix"}. The time series in row \code{entity_name} and column \code{field_name} contains all observations of such entity for such field.

The number of rows is given by to the number of distinct entity names (parameter \code{names}), while the number of columns is given by the number of distinct attribute/field names (parameter \code{fields}).
}
\description{
Create a \code{"uts_matrix"} from \emph{long} (also known as \emph{narrow}) tabular data. Data in this format has four different columns; the observation values, the observation times, the entity name of each observation (e.g. person, country), and the attribute/field name of each observation (e.g. which one of serveral economic indicators, which one of several blood measurement values).
}
\examples{
values <- c(A=1, A=2, B=3, B=4, A=5)
times <- as.POSIXct("2016-01-01") + dhours(1:5)
uts_matrix_long(values, times, fields=c("c", "d", "d", "d", "d"))


uts_matrix_long(values=1:5, times=as.POSIXct("2015-01-01") + days(1:5),
  names=c("A", "A", "B", "B", "A"), fields=c("c", "d", "d", "d", "d"))
}
\seealso{
\code{\link{uts_matrix_wide}}, \code{\link{uts_vector_long}}, \code{\link{uts_vector_wide}}
}
\keyword{classes}
\keyword{ts}

