% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_matrix.R
\name{uts_matrix_long}
\alias{uts_matrix_long}
\title{Create uts_matrix from long tabular data}
\usage{
uts_matrix_long(values, times, rownames, colnames)
}
\arguments{
\item{values}{a vector observation values.}

\item{times}{a \code{\link{POSIXct}} object. The matching observation times.}

\item{rownames}{a character vector. The the row name (also known as \emph{record}) of each observation.}

\item{colnames}{a character vector. The the column names (also known as \emph{field}) of each observation.}
}
\value{
An object of class \code{"uts_matrix"}. The number of rows is given by to the number of records (distinct \code{rownames}), while the number of columns is given by the number of fields (distinct \code{colnames}).
}
\description{
Create a \code{"uts_matrix"} from \emph{long} (also known as \emph{narrow}) tabular data. Data in this format has four different columns; the observation values, the observation times, and the row and column name (also known as \emph{record} and \emph{field}) of each observation.
}
\examples{
uts_matrix_long(values=1:5, times=as.POSIXct("2015-01-01") + days(1:5),
  colnames=c("A", "A", "B", "B", "A"), rownames=c("c", "d", "d", "d", "d"))
}
\keyword{classes}
\keyword{ts}

