% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{uts_vector}
\alias{c.uts}
\alias{c.uts_vector}
\alias{is.uts_vector}
\alias{uts_vector}
\title{Unevenly-spaced Time Series Vector}
\usage{
uts_vector(...)

\method{c}{uts}(...)

\method{c}{uts_vector}(...)

is.uts_vector(x)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{zero or more \code{\link{uts}} objects.}
}
\value{
An object of class \code{"uts_vector"}.
}
\description{
Create a vector of unevenly spaced time series (\code{"uts_vector"}).

\code{is.uts_vector} returns \code{TRUE} if its argument is a \code{"uts_vector"} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{c}: constructor for \code{"uts_vector"} object out of \code{"uts"} objects.

\item \code{c}: constructor for \code{"uts_vector"} object out of \code{"uts"} and other \code{"uts_vector"} objects.
}}
\note{
A pure virtual (abstract) class \code{"uts_virtual"} exists from which \code{"uts"}, \code{"uts_vector"}, \code{"uts_matrix"}, and \code{"uts_data_frame"} inherit: it is used to allow operations such as subtraction to mix the classes.
}
\examples{
# Two equivalent constructors
uts_vector(apples=ex_uts(), oranges=ex_uts2())
c(apples=ex_uts(), oranges=ex_uts2())

# Create "uts_vector" out of both "uts" and "uts_vector" objects
c(ex_uts(), ex_uts_vector(), kiwis=ex_uts2())

# Empty "uts_vector"
uts_vector()

# The first two test return TRUE, the other returns FALSE
is.uts_vector(uts_vector())
is.uts_matrix(uts_matrix())      # "uts_matrix" inherits from "uts_vector"
is.uts_vector(ex_uts())
}
\seealso{
\code{\link{rep.uts}}, \code{\link{uts_vector_long}}, \code{\link{uts_vector_wide}} for alternative constructors.
}
\keyword{classes}
\keyword{internal}
\keyword{ts}

