% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts_vector.R
\name{uts_vector_wide}
\alias{uts_vector_wide}
\title{Create uts_vector from wide tabular data}
\usage{
uts_vector_wide(values, times, names = colnames(values))
}
\arguments{
\item{values}{a matrix or data.frame. Each row represents a vector of observations at a specific time point.}

\item{times}{a \code{\link{POSIXct}} object. The observation times of the rows of \code{values}.}

\item{names}{a character vector, indicating the source of each column in \code{values}. By default, the column names of \code{values} are used.}
}
\value{
An object of class \code{"uts_vector"}. The number of time series is equal to the number of columns of \code{values}. The length of each time series is equal to the number of rows of \code{values}.
}
\description{
Create a \code{"uts_vector"} from wide tabular data (see \href{https://en.wikipedia.org/wiki/Wide_and_narrow_data}{Wikipedia}). With this input format, values in the same row are measurements at the same point in time, while values in the same column are measurements of the same variable.
}
\examples{
data <- data.frame(apples=1:10, oranges=letters[1:10], bananas=month.name[1:10])
uts_vector_wide(data, times=as.POSIXct("2015-01-01") + ddays(1:10))
}
\keyword{classes}
\keyword{ts}

