% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utsMultivariate-internal}
\alias{utsMultivariate-internal}
\title{Internal Functions}
\description{
The internal functions listed below might be of interest to developers seeking to extend the package functionality.
}
\details{
\code{uts_vector} methods:
\itemize{
  \item \code{\link{is.uts_matrix}}
  \item \code{\link{is.uts_vector}}
  \item \code{\link{print.uts_matrix}}
  \item \code{\link{print.uts_vector}}
}

Helper functions:
\itemize{
  \item \code{\link{Ops_uts_vector}}
}

Methods that exists solely ensure that methods intended for \code{\link{ts}} objects in base \R are not accidentally applied to \code{"uts_vector"} and \code{"uts_matrix"} objects:
\itemize{
  \item \code{\link{as.ts.uts_vector}}
  \item \code{\link{cycle.uts_vector}}
  \item \code{\link{frequency.uts_vector}}
}
}

